Adding the ability to record sim input events and animation frames, so they can be precisely played back to duplicate performance characteristics.
Adds two options to the menu shown when 'recordInputEventLog' query parameter exists: one to print log information to the console (call sim.startInputEventPlayback with this),
or one to upload it to a running Node.js server (scenery/tests/event-logs/server/server.js).

When the 'playbackInputEventLog' query parameter is given, it will load input events and animation frames from the Node.js server. It loads the last saved data from the
unique combination of browser (user agent string), window.innerWidth, and window.innerHeight, so different data can be used on different devices/browsers.
This is necessary for now because of coordinates changing depending on the size of the sim, and possibly other different behavior.